rpc,vendor: add openrpc.json as bundled static resource
This is just a proof of concept.

Up for grabs would be using other bundlers, putting the
openrpc.json file somewhere else.

Probably the packr-generated (or similar) file should be built as a step
of the Make process or similar.

As-is, current steps ar:
- packr
- make geth

Note that I had to manually go thru and use

govendor add ___

to slog thru adding missing deps for packr. If we change the bundler package,
that's one way of doing that (but hopefully not the only or best).